Migrate mortgage calculator code to seperate component
